requirements 1.txt
mysql-connector-python==9.3.0
python-dotenv==1.1.0
pandas==2.2.3
plotly==6.0.1
nbformat==5.10.4
openpyxl==3.1.5